tools/accident_collector/requirements.txt
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0.post0
pydantic==2.11.7
tenacity==9.0.0
pdfminer.six==20231228
